Add an initial prototype mobile experience.
Kinda blows to bring in the whole of Sencha UI just to get its carousel component, but they seem to have the only decent swipe-to-switch implementation.

Maybe we can build a subset of Sencha UI that only includes these parts, but I get the feeling that this framework is pretty intertwingled.