Implement a sliding window of items in mobile view.
This presents the illusion of an infinite stream of items while constraining the number of items in memory at once.

There is still a crash bug on my Samsung Epic but I'm beginning to think that's the Flash plugin crashing since it doesn't doesn't recreate on an emulator without Flash available.